LPS-106487 Move resize logic from Column to ColumnWithControls
