+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/app/components/layout-data-items/Column.js
@@ -12,14 +12,11 @@
  * details.
  */
 
-import ClayButton from '@clayui/button';
 import classNames from 'classnames';
-import React, {useContext, useMemo} from 'react';
+import React from 'react';
 
 import {LAYOUT_DATA_ITEM_DEFAULT_CONFIGURATIONS} from '../../config/constants/layoutDataItemDefaultConfigurations';
 import {LAYOUT_DATA_ITEM_TYPES} from '../../config/constants/layoutDataItemTypes';
-import {useSelector} from '../../store/index';
-import {ResizingContext} from './RowWithControls';
 
 function Column({children, className, item, ...props}, ref) {
 	const {
@@ -30,47 +27,6 @@ function Column({children, className, item, ...props}, ref) {
 		}
 	} = item;
 
-	const layoutData = useSelector(state => state.layoutData);
-
-	const {onResizeEnd, onResizeStart, onResizing} = useContext(
-		ResizingContext
-	);
-
-	const columnInfo = useMemo(() => getColumnInfo({item, layoutData}), [
-		item,
-		layoutData
-	]);
-
-	const onMouseMoveInWindow = event => {
-		event.preventDefault();
-
-		onResizing(event, columnInfo);
-	};
-
-	const onMouseDownInWindow = event => {
-		onResizeStart(event);
-
-		window.addEventListener('mouseup', onMouseUpInWindow);
-		window.addEventListener('mousemove', onMouseMoveInWindow);
-	};
-
-	const onMouseUpInWindow = event => {
-		onResizeEnd(event);
-
-		window.removeEventListener('mouseup', onMouseUpInWindow);
-		window.removeEventListener('mousemove', onMouseMoveInWindow);
-	};
-
-	const resizeHandler = (
-		<div>
-			{children}
-			<ClayButton
-				className="page-editor__col-resizer"
-				onMouseDown={onMouseDownInWindow}
-			/>
-		</div>
-	);
-
 	return (
 		<>
 			<div
@@ -80,43 +36,10 @@ function Column({children, className, item, ...props}, ref) {
 				ref={ref}
 				{...props}
 			>
-				{!columnInfo.isLastColumn ? resizeHandler : children}
+				{children}
 			</div>
 		</>
 	);
 }
 
 export default React.forwardRef(Column);
-
-/**
- * Retrieves necessary data from the current and next column.
- *
- * @param {!Object} options
- * @param {!Object} options.item
- * @param {!Object} options.layoutData
- *
- * @returns {!Object}
- */
-function getColumnInfo({item, layoutData}) {
-	const rowColumns = layoutData.items[item.parentId].children;
-	const colIndex = rowColumns.indexOf(item.itemId);
-	const nextColumnIndex = colIndex + 1;
-	const currentColumn = item;
-	const currentColumnConfig = currentColumn.config;
-	const nextColumn = {...layoutData.items[rowColumns[nextColumnIndex]]};
-	const nextColumnConfig =
-		typeof nextColumn === 'object' && Object.keys(nextColumn).length
-			? nextColumn.config
-			: {};
-
-	return {
-		colIndex,
-		currentColumn,
-		currentColumnConfig,
-		isLastColumn: rowColumns.indexOf(item.itemId) === rowColumns.length - 1,
-		nextColumn: nextColumn ? nextColumn : {},
-		nextColumnConfig: nextColumn ? nextColumnConfig : {},
-		nextColumnIndex: colIndex + 1,
-		rowColumns
-	};
-}
